Use FST serializer for all objects (not only encrypted ones)
#87

+++ org.hive2hive.core/src/test/java/org/hive2hive/core/processes/files/add/AddFileTest.java
@@ -163,8 +163,12 @@ public class AddFileTest extends H2HJUnitTest {
 	@AfterClass
 	public static void endTest() throws IOException {
 		NetworkTestUtil.shutdownNetwork(network);
-		FileUtils.deleteDirectory(uploaderRoot);
-		FileUtils.deleteDirectory(downloaderRoot);
+		if (uploaderRoot != null && uploaderRoot.exists()) {
+			FileUtils.deleteDirectory(uploaderRoot);
+		}
+		if (downloaderRoot != null && downloaderRoot.exists()) {
+			FileUtils.deleteDirectory(downloaderRoot);
+		}
 		afterClass();
 	}
 }
